# apps/backend-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APIResponse:
    properties:
      count:
        type: integer
      data: {}
      error:
        type: string
      has_more:
        type: boolean
      next_cursor:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    name: gomo6 Team
    url: https://gomo6.wtf
  description: gomo6 — open social platform API. Create bots, integrations, and apps.
  license:
    name: MIT
  termsOfService: https://gomo6.wtf/terms
  title: gomo6 API
  version: "1.0"
paths:
...
  /threads:
    delete:
      description: Delete a thread (author only)
      parameters:
      - description: Thread ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete thread
      tags:
      - Threads
    get:
      description: Get threads with optional filters
      parameters:
      - description: Filter by board ID
        in: query
        name: board_id
        type: string
      - description: Filter by channel ID
        in: query
        name: channel_id
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - default: 50
        description: Max results (1-100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: List threads
      tags:
      - Threads
  /threads/{id}:
    get:
      description: Get a thread by its ID
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get thread
      tags:
      - Threads
    put:
      consumes:
      - application/json
      description: Update thread content (author only)
      parameters:
      - description: Thread ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update thread
      tags:
      - Threads
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Bearer token for API authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"